Creating new hook for onboarding flow
