+++ src/views/BookkeepingOverview/BookkeepingOverview.tsx
@@ -1,14 +1,8 @@
-import { useCallback, useEffect, useState } from 'react'
 import classNames from 'classnames'
 import { PopupModal } from 'react-calendly'
 import { useTranslation } from 'react-i18next'
 
-import { type CallBooking as CallBookingData, CallBookingPurpose, CallBookingType } from '@schemas/callBooking'
 import { type Variants } from '@utils/styleUtils/sizeVariants'
-import { useBookkeepingStatus, useBookkeepingStatusGlobalCacheActions } from '@hooks/api/businesses/[business-id]/bookkeeping/status/useBookkeepingStatus'
-import { useCallBookings } from '@hooks/api/businesses/[business-id]/call-bookings/useCallBookings'
-import { useCreateCallBooking } from '@hooks/api/businesses/[business-id]/call-bookings/useCreateCallBooking'
-import { useCalendly } from '@hooks/features/calendly/useCalendly'
 import { useSizeClass, useWindowSize } from '@hooks/utils/size/useWindowSize'
 import { VStack } from '@ui/Stack/Stack'
 import { CallBooking } from '@components/CallBooking/CallBooking'
@@ -23,6 +17,7 @@ import { ProfitAndLossOverviewDetailedCharts } from '@components/ProfitAndLossOv
 import { type ProfitAndLossSummariesStringOverrides } from '@components/ProfitAndLossSummaries/ProfitAndLossSummaries'
 import { Tasks, type TasksStringOverrides } from '@components/Tasks/Tasks'
 import { View } from '@components/View/View'
+import { useBookkeepingOverviewCalendly } from '@views/BookkeepingOverview/useBookkeepingOverviewCalendly'
 import { useKeepInMobileViewport } from '@views/BookkeepingOverview/useKeepInMobileViewport'
 
 export interface BookkeepingOverviewProps {
@@ -68,103 +63,15 @@ export const BookkeepingOverview = ({
   const { upperContentRef, targetElementRef, upperElementInFocus } =
     useKeepInMobileViewport()
 
-  const { data: bookkeepingStatus } = useBookkeepingStatus()
-  const { forceReloadBookkeepingStatus } = useBookkeepingStatusGlobalCacheActions()
-  const { trigger: createCallBooking } = useCreateCallBooking()
-
-  const [embedDismissed, setEmbedDismissed] = useState(false)
-  const [hasScheduledCallInSession, setHasScheduledCallInSession] = useState(false)
-
-  const onboardingCallUrlFromApi =
-    bookkeepingStatus != null
-    && bookkeepingStatus.showEmbeddedOnboarding
-    && bookkeepingStatus.onboardingCallUrl != null
-      ? bookkeepingStatus.onboardingCallUrl
-      : undefined
-
-  const recordCalendlyScheduled = useCallback(async (payload?: { event: { uri: string } }) => {
-    const eventUri = payload?.event.uri
-
-    let externalId: string | undefined
-    if (eventUri) {
-      try {
-        const segments = new URL(eventUri).pathname.split('/').filter(Boolean)
-        externalId = segments.at(-1)
-      }
-      catch (_) {
-        externalId = undefined
-      }
-    }
-
-    if (externalId) {
-      setHasScheduledCallInSession(true)
-
-      try {
-        await createCallBooking({
-          external_id: externalId,
-          purpose: CallBookingPurpose.BOOKKEEPING_ONBOARDING,
-          call_type: CallBookingType.GOOGLE_MEET,
-        })
-        void forceReloadBookkeepingStatus()
-      }
-      catch (error: unknown) {
-        console.error('Failed to record onboarding call booking', error)
-      }
-    }
-  }, [createCallBooking, forceReloadBookkeepingStatus])
-
-  const handleCalendlyScheduled = useCallback((payload?: { event: { uri: string } }) => {
-    void recordCalendlyScheduled(payload)
-  }, [recordCalendlyScheduled])
-
-  const { isCalendlyVisible, calendlyLink, calendlyRef, openCalendly, closeCalendly } = useCalendly({
-    onEventScheduled: handleCalendlyScheduled,
-  })
-
-  const { data: callBookings, isError, isLoading, isValidating } = useCallBookings({ limit: 1 })
-  const callBooking: CallBookingData | null = callBookings?.[0]?.data[0] ?? null
-  const hasResolvedCallBooking = !isLoading && !isValidating && !isError
-  const callBookingVisible = hasResolvedCallBooking && callBooking != null
-  const callBookingEmptyStateVisible = hasResolvedCallBooking
-    && callBooking == null
-    && onboardingCallUrlFromApi != null
-    && embedDismissed
-    && !hasScheduledCallInSession
-
-  const handleBookCall = useCallback(() => {
-    if (onboardingCallUrlFromApi != null) {
-      openCalendly(onboardingCallUrlFromApi)
-    }
-  }, [onboardingCallUrlFromApi, openCalendly])
-
-  const handleCloseCalendly = useCallback(() => {
-    setEmbedDismissed(true)
-    void forceReloadBookkeepingStatus()
-    closeCalendly()
-  }, [closeCalendly, forceReloadBookkeepingStatus])
-
-  useEffect(() => {
-    if (
-      !hasResolvedCallBooking
-      || callBooking != null
-      || onboardingCallUrlFromApi == null
-      || embedDismissed
-      || hasScheduledCallInSession
-      || isCalendlyVisible
-    ) {
-      return
-    }
-
-    openCalendly(onboardingCallUrlFromApi)
-  }, [
+  const {
     callBooking,
-    embedDismissed,
-    hasScheduledCallInSession,
-    hasResolvedCallBooking,
+    showCallBookingCard,
+    handleBookCall,
     isCalendlyVisible,
-    onboardingCallUrlFromApi,
-    openCalendly,
-  ])
+    calendlyLink,
+    calendlyRef,
+    handleCloseCalendly,
+  } = useBookkeepingOverviewCalendly()
 
   return (
     <ProfitAndLoss asContainer={false}>
@@ -183,9 +90,9 @@ export const BookkeepingOverview = ({
         withSidebar={width > 1100}
         sidebar={(
           <VStack gap='lg'>
-            {(callBookingVisible || callBookingEmptyStateVisible) && (
+            {showCallBookingCard && (
               <CallBooking
-                callBooking={callBooking ?? undefined}
+                callBooking={callBooking}
                 onBookCall={handleBookCall}
               />
             )}
@@ -203,9 +110,9 @@ export const BookkeepingOverview = ({
             onClick={() => (upperElementInFocus.current = true)}
           >
             <VStack gap='lg'>
-              {(callBookingVisible || callBookingEmptyStateVisible) && (
+              {showCallBookingCard && (
                 <CallBooking
-                  callBooking={callBooking ?? undefined}
+                  callBooking={callBooking}
                   onBookCall={handleBookCall}
                 />
               )}
